requirements.txt
flask==3.1.2
google-genai==1.67.0
requests==2.32.5
